requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
h11==0.14.0
idna==3.4
numpy==1.23.3
outcome==1.2.0
pandas==1.5.0
PySocks==1.7.1
pyTelegramBotAPI==4.7.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.4
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
wsproto==1.2.0
